hoststorage/requirements.txt
pyyaml
git+git://github.com/shotgunsoftware/python-api.git@v3.2.2
